+++ src/components/Header.tsx
@@ -1,7 +1,6 @@
 import styled from "styled-components";
 
 const HeaderStyle = styled.header`
-  position: fixed;
   height: 40px;
   width: 100%;
   display: flex;
