+++ src/pages/_app.tsx
@@ -1,5 +1,6 @@
 import type { AppProps } from "next/app";
 import { ThemeProvider } from "styled-components";
+import Layout from "../components/Layout";
 
 import { GlobalStyle } from "../styles/global-style";
 import { theme } from "../styles/theme";
